Related work items: #16917

+++ src/main/java/com/mashreq/transfercoreservice/fundtransfer/service/QRDealsService.java
@@ -2,9 +2,11 @@ package com.mashreq.transfercoreservice.fundtransfer.service;
 
 import com.mashreq.transfercoreservice.fundtransfer.dto.QRDealDetails;
 
+import java.math.BigDecimal;
+
 public interface QRDealsService {
 
     QRDealDetails getQRDealDetails(String cif, String country);
 
-    void updateQRDeals(String cif, Float utilizedAmount);
+    void updateQRDeals(String cif, BigDecimal utilizedAmount);
 }
